Update __init__.py and Add new file ../hdfs/utils/__init__.py
